Split routes out from server.ts
